Limit path import for past days
Inside path_etl(), limit the path reimport of past days to 30 days.
Otherwise, all scooter IDs get loaded and this takes a lot of time.
Downside: If a scooter with the same ID e.g. gets repaired and then
released again, it will show twice with a NULL point_before.