Prepared DAG for multiple cities
Instead of just calculating paths for Winterthur, prepared the DAG for
calculating paths for multiple cities in Switzerland. The config of the
different cities are stored in module `sharedmobility.config`.

Added city_bfs_id to persistance in database tables.
Some cities do not have data from start_date of the DAG.